Fix type problem with stdarg interface to dialog_box.
git-svn-id: svn://svn.berlios.de/lincity-ng/trunk@30 8226ee30-1aee-0310-a633-b68f58d1eb45

+++ dialbox.c
@@ -72,7 +72,7 @@ dialog_box(int arg_color, char argc, ...)
   for (i = 0; i < argc; i++) {
     dbn++;
 
-    dbox_entry[i].type = va_arg(ap, short);
+    dbox_entry[i].type = (short) va_arg(ap, int);
     dbox_entry[i].text = va_arg(ap, char *);
 
     if (dbox_entry[i].type) {
